+++ packages/deploy-aws/src/seagull_app.ts
@@ -1,15 +1,11 @@
 import { Role } from '@aws-cdk/aws-iam'
-import { App, StackProps } from '@aws-cdk/cdk'
+import { App } from '@aws-cdk/cdk'
 import * as cfnDiff from '@aws-cdk/cloudformation-diff'
 import * as cdk from 'aws-cdk'
 import * as yaml from 'yaml'
 import * as lib from './lib'
 import { SeagullStack } from './seagull_stack'
-
-interface OperationsProps {
-  projectName: string
-  stackProps?: StackProps
-}
+import { OperationsProps } from './types'
 
 export class SeagullApp extends App {
   stack: SeagullStack
